+++ axelor-account/src/main/java/com/axelor/apps/account/service/payment/invoice/payment/InvoicePaymentCreateService.java
@@ -40,7 +40,8 @@ public interface InvoicePaymentCreateService {
       PaymentMode paymentMode,
       int typeSelect);
 
-  public InvoicePayment createInvoicePayment(Invoice invoice, BigDecimal amount, Move paymentMove)
+  public InvoicePayment createInvoicePayment(
+      Invoice invoice, BigDecimal amount, Move paymentMove, LocalDate paymentDate)
       throws AxelorException;
 
   /**